+++ CHANGELOG.md
@@ -1,15 +1,24 @@
 # Changelog
-### v0.0.51
 
-- Checkbox - Modify to Wrap Text 
+### v0.0.52
+
+- Switch - Add new component
 
 --
+
+### v0.0.51
+
+- Checkbox - Modify to Wrap Text
+
+--
+
 ### v0.0.50
 
-- Filter Drawer - Added prop to make header sticky for secondary screen 
+- Filter Drawer - Added prop to make header sticky for secondary screen
 - Filter Drawer - Added Layout prop to get layout information for secondary screen
 
 --
+
 ### v0.0.49
 
 - Group list item - Fix font weight when inversed